feat: move 6 portable screen/loggedIn files to commons for KMP
Moved to commons:
- video/datasource/FeedBasis.kt (video feed mime types and kind constants)
- video/dal/SupportedContent.kt (URL content type validator)
- chess/datasource/FilterChessEvent.kt (chess relay filter builder)
- chess/datasource/ChessQueryState.kt (extracted from SubAssembler)
- geohash/datasource/GeohashQueryState.kt (extracted from FilterAssembler)
- hashtag/datasource/HashtagQueryState.kt (extracted from FilterAssembler)

Updated amethyst originals with typealias/re-exports and commons imports.
Part of KMP iOS migration tracked in #2238.
