+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/hashtag/datasource/HashtagFilterAssembler.kt
@@ -21,16 +21,8 @@
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.hashtag.datasource
 
 import com.vitorpamplona.amethyst.commons.relayClient.composeSubscriptionManagers.ComposeSubscriptionManager
+import com.vitorpamplona.amethyst.commons.ui.screen.loggedIn.hashtag.datasource.HashtagQueryState
 import com.vitorpamplona.quartz.nip01Core.relay.client.INostrClient
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-
-// This allows multiple screen to be listening to tags, even the same tag
-class HashtagQueryState(
-    val hashtag: String,
-    val relays: Set<NormalizedRelayUrl>,
-) {
-    val lowercaseHashtag = hashtag.lowercase()
-}
 
 class HashtagFilterAssembler(
     client: INostrClient,
